# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Conserje API
  version: 2.0.0
  description: Plataforma completa de gestão condominial
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visitantes:
    get:
      operationId: listVisitantes
      tags: [visitantes]
      parameters:
        - name: tipo
          in: query
          schema: { type: string }
        - name: data
          in: query
          schema: { type: string }
        - name: dentro
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visitante"
    post:
      operationId: createVisitante
      tags: [visitantes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitanteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitante"
  /visitantes/{id}/saida:
    post:
      operationId: registrarSaida
      tags: [visitantes]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitante"

  # ─── OCORRÊNCIAS ─────────────────────────────────────────────────────────────
  /ocorrencias:
    get:
      operationId: listOcorrencias
      tags: [ocorrencias]
      parameters:
        - name: tipo
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: prioridade
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ocorrencia"
    post:
      operationId: createOcorrencia
      tags: [ocorrencias]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcorrenciaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ocorrencia"
  /ocorrencias/{id}:
    patch:
      operationId: updateOcorrencia
      tags: [ocorrencias]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcorrenciaInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ocorrencia"

  # ─── FINANCEIRO ──────────────────────────────────────────────────────────────
  /financeiro:
    get:
      operationId: listFinanceiro
      tags: [financeiro]
      parameters:
        - name: tipo
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: categoria
          in: query
          schema: { type: string }
        - name: mes
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Financeiro"
    post:
      operationId: createFinanceiro
      tags: [financeiro]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceiroInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Financeiro"
  /financeiro/{id}:
    patch:
      operationId: updateFinanceiro
      tags: [financeiro]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceiroInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Financeiro"
...
  /assembleias:
    get:
      operationId: listAssembleias
      tags: [assembleias]
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assembleia"
    post:
      operationId: createAssembleia
      tags: [assembleias]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssembleiaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assembleia"
  /assembleias/{id}:
    get:
      operationId: getAssembleia
      tags: [assembleias]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssembleiaComVotacoes"
    patch:
      operationId: updateAssembleia
      tags: [assembleias]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssembleiaInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assembleia"

  # ─── RESERVAS ────────────────────────────────────────────────────────────────
...
  /reservas:
    get:
      operationId: listReservas
      tags: [reservas]
      parameters:
        - name: area_id
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: data
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReservaComDetalhes"
    post:
      operationId: createReserva
      tags: [reservas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaComDetalhes"
  /reservas/{id}:
    patch:
      operationId: updateReserva
      tags: [reservas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservaInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaComDetalhes"

  # ─── JURÍDICO ────────────────────────────────────────────────────────────────
  /juridico/documentos:
    get:
      operationId: listDocumentos
      tags: [juridico]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Documento"
    post:
      operationId: createDocumento
      tags: [juridico]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Documento"
  /juridico/notificacoes:
    get:
      operationId: listNotificacoesJuridicas
      tags: [juridico]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: tipo
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificacaoJuridica"
    post:
      operationId: createNotificacaoJuridica
      tags: [juridico]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificacaoJuridicaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificacaoJuridica"
  /juridico/notificacoes/{id}:
    patch:
      operationId: updateNotificacaoJuridica
      tags: [juridico]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificacaoJuridicaInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificacaoJuridica"

  # ─── ALERTAS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Visitante:
      type: object
      required: [id, nome, destino_apartamento, destino_bloco, tipo, entrada, created_at]
      properties:
        id: { type: integer }
        nome: { type: string }
        documento: { type: string, nullable: true }
        telefone: { type: string, nullable: true }
        foto_url: { type: string, nullable: true }
        destino_apartamento: { type: string }
        destino_bloco: { type: string }
        destino_morador: { type: string, nullable: true }
        motivo: { type: string }
        tipo: { type: string }
        veiculo_placa: { type: string, nullable: true }
        qr_code: { type: string, nullable: true }
        entrada: { type: string }
        saida: { type: string, nullable: true }
        autorizado_por: { type: string, nullable: true }
        observacoes: { type: string, nullable: true }
        created_at: { type: string }
    VisitanteInput:
      type: object
      required: [nome, destino_apartamento, destino_bloco]
      properties:
        nome: { type: string }
        documento: { type: string }
        telefone: { type: string }
        destino_apartamento: { type: string }
        destino_bloco: { type: string }
        destino_morador: { type: string }
        motivo: { type: string }
        tipo: { type: string }
        veiculo_placa: { type: string }
        autorizado_por: { type: string }
        observacoes: { type: string }
    Ocorrencia:
      type: object
      required: [id, titulo, descricao, tipo, prioridade, status, created_at]
      properties:
        id: { type: integer }
        titulo: { type: string }
        descricao: { type: string }
        tipo: { type: string }
        prioridade: { type: string }
        status: { type: string }
        morador_id: { type: integer, nullable: true }
        area: { type: string, nullable: true }
        foto_url: { type: string, nullable: true }
        prestador_responsavel: { type: string, nullable: true }
        valor_servico: { type: string, nullable: true }
        data_agendamento: { type: string, nullable: true }
        data_conclusao: { type: string, nullable: true }
        observacoes: { type: string, nullable: true }
        created_at: { type: string }
        updated_at: { type: string, nullable: true }
    OcorrenciaInput:
      type: object
      required: [titulo, descricao]
      properties:
        titulo: { type: string }
        descricao: { type: string }
        tipo: { type: string }
        prioridade: { type: string }
        status: { type: string }
        morador_id: { type: integer }
        area: { type: string }
        prestador_responsavel: { type: string }
        valor_servico: { type: string }
        data_agendamento: { type: string }
        observacoes: { type: string }
    Financeiro:
      type: object
      required: [id, tipo, descricao, valor, categoria, status, created_at]
      properties:
        id: { type: integer }
        tipo: { type: string }
        descricao: { type: string }
        valor: { type: number }
        categoria: { type: string }
        data_vencimento: { type: string, nullable: true }
        data_pagamento: { type: string, nullable: true }
        morador_id: { type: integer, nullable: true }
        referencia_mes: { type: string, nullable: true }
        status: { type: string }
        comprovante_url: { type: string, nullable: true }
        observacoes: { type: string, nullable: true }
        created_at: { type: string }
        updated_at: { type: string, nullable: true }
    FinanceiroInput:
      type: object
      required: [tipo, descricao, valor]
      properties:
        tipo: { type: string }
        descricao: { type: string }
        valor: { type: number }
        categoria: { type: string }
        data_vencimento: { type: string }
        data_pagamento: { type: string }
        morador_id: { type: integer }
        referencia_mes: { type: string }
        status: { type: string }
        observacoes: { type: string }
...
    Assembleia:
      type: object
      required: [id, titulo, tipo, data_realizacao, local, status, created_at]
      properties:
        id: { type: integer }
        titulo: { type: string }
        tipo: { type: string }
        descricao: { type: string, nullable: true }
        data_realizacao: { type: string }
        local: { type: string }
        link_reuniao: { type: string, nullable: true }
        status: { type: string }
        quorum_minimo: { type: integer, nullable: true }
        presentes: { type: integer, nullable: true }
        ata_url: { type: string, nullable: true }
        ata_texto: { type: string, nullable: true }
        created_at: { type: string }
        updated_at: { type: string, nullable: true }
    AssembleiaComVotacoes:
      allOf:
        - $ref: "#/components/schemas/Assembleia"
        - type: object
          properties:
            votacoes:
              type: array
              items:
                $ref: "#/components/schemas/Votacao"
    Votacao:
      type: object
      required: [id, assembleia_id, questao, status, created_at]
      properties:
        id: { type: integer }
        assembleia_id: { type: integer }
        questao: { type: string }
        descricao: { type: string, nullable: true }
        opcoes: { type: string }
        votos_sim: { type: integer, nullable: true }
        votos_nao: { type: integer, nullable: true }
        votos_abstencao: { type: integer, nullable: true }
        status: { type: string }
        resultado: { type: string, nullable: true }
        created_at: { type: string }
    AssembleiaInput:
      type: object
      required: [titulo, data_realizacao]
      properties:
        titulo: { type: string }
        tipo: { type: string }
        descricao: { type: string }
        data_realizacao: { type: string }
        local: { type: string }
        link_reuniao: { type: string }
        status: { type: string }
        quorum_minimo: { type: integer }
        ata_texto: { type: string }
...
    ReservaComDetalhes:
      type: object
      required: [id, area_id, morador_id, data_inicio, data_fim, status, created_at]
      properties:
        id: { type: integer }
        area_id: { type: integer }
        morador_id: { type: integer }
        data_inicio: { type: string }
        data_fim: { type: string }
        status: { type: string }
        motivo: { type: string, nullable: true }
        num_pessoas: { type: integer, nullable: true }
        valor_cobrado: { type: string, nullable: true }
        observacoes: { type: string, nullable: true }
        area_nome: { type: string, nullable: true }
        morador_nome: { type: string, nullable: true }
        morador_apartamento: { type: string, nullable: true }
        created_at: { type: string }
    ReservaInput:
      type: object
      required: [area_id, morador_id, data_inicio, data_fim]
      properties:
        area_id: { type: integer }
        morador_id: { type: integer }
        data_inicio: { type: string }
        data_fim: { type: string }
        status: { type: string }
        motivo: { type: string }
        num_pessoas: { type: integer }
        observacoes: { type: string }
    Documento:
      type: object
      required: [id, titulo, tipo, vigente, created_at]
      properties:
        id: { type: integer }
        titulo: { type: string }
        tipo: { type: string }
        descricao: { type: string, nullable: true }
        arquivo_url: { type: string, nullable: true }
        versao: { type: string, nullable: true }
        vigente: { type: string }
        created_at: { type: string }
        updated_at: { type: string, nullable: true }
    DocumentoInput:
      type: object
      required: [titulo]
      properties:
        titulo: { type: string }
        tipo: { type: string }
        descricao: { type: string }
        arquivo_url: { type: string }
        versao: { type: string }
        vigente: { type: string }
    NotificacaoJuridica:
      type: object
      required: [id, tipo, titulo, descricao, status, created_at]
      properties:
        id: { type: integer }
        morador_id: { type: integer, nullable: true }
        tipo: { type: string }
        titulo: { type: string }
        descricao: { type: string }
        valor_multa: { type: number, nullable: true }
        status: { type: string }
        data_prazo: { type: string, nullable: true }
        data_resolucao: { type: string, nullable: true }
        artigo_regimento: { type: string, nullable: true }
        comprovante_url: { type: string, nullable: true }
        created_at: { type: string }
        morador_nome: { type: string, nullable: true }
        morador_apartamento: { type: string, nullable: true }
    NotificacaoJuridicaInput:
      type: object
      required: [tipo, titulo, descricao]
      properties:
        morador_id: { type: integer }
        tipo: { type: string }
        titulo: { type: string }
        descricao: { type: string }
        valor_multa: { type: number }
        status: { type: string }
        data_prazo: { type: string }
        artigo_regimento: { type: string }
...