# lib/api-zod/src/generated/types/assembleiaComVotacoes.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plataforma completa de gestão condominial
 * OpenAPI spec version: 2.0.0
 */
import type { Assembleia } from "./assembleia";
import type { Votacao } from "./votacao";

export type AssembleiaComVotacoes = Assembleia & {
  votacoes?: Votacao[];
};