# lib/api-zod/src/generated/types/assembleiaInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plataforma completa de gestão condominial
 * OpenAPI spec version: 2.0.0
 */

export interface AssembleiaInput {
  titulo: string;
  tipo?: string;
  descricao?: string;
  data_realizacao: string;
  local?: string;
  link_reuniao?: string;
  status?: string;
  quorum_minimo?: number;
  ata_texto?: string;
}