# lib/api-zod/src/generated/types/financeiro.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plataforma completa de gestão condominial
 * OpenAPI spec version: 2.0.0
 */

export interface Financeiro {
  id: number;
  tipo: string;
  descricao: string;
  valor: number;
  categoria: string;
  data_vencimento?: string | null;
  data_pagamento?: string | null;
  morador_id?: number | null;
  referencia_mes?: string | null;
  status: string;
  comprovante_url?: string | null;
  observacoes?: string | null;
  created_at: string;
  updated_at?: string | null;
}