# lib/api-zod/src/generated/types/reservaComDetalhes.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plataforma completa de gestão condominial
 * OpenAPI spec version: 2.0.0
 */

export interface ReservaComDetalhes {
  id: number;
  area_id: number;
  morador_id: number;
  data_inicio: string;
  data_fim: string;
  status: string;
  motivo?: string | null;
  num_pessoas?: number | null;
  valor_cobrado?: string | null;
  observacoes?: string | null;
  area_nome?: string | null;
  morador_nome?: string | null;
  morador_apartamento?: string | null;
  created_at: string;
}