# lib/api-zod/src/generated/types/reservaInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plataforma completa de gestão condominial
 * OpenAPI spec version: 2.0.0
 */

export interface ReservaInput {
  area_id: number;
  morador_id: number;
  data_inicio: string;
  data_fim: string;
  status?: string;
  motivo?: string;
  num_pessoas?: number;
  observacoes?: string;
}