# lib/api-zod/src/generated/types/votacao.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Plataforma completa de gestão condominial
 * OpenAPI spec version: 2.0.0
 */

export interface Votacao {
  id: number;
  assembleia_id: number;
  questao: string;
  descricao?: string | null;
  opcoes?: string;
  votos_sim?: number | null;
  votos_nao?: number | null;
  votos_abstencao?: number | null;
  status: string;
  resultado?: string | null;
  created_at: string;
}